Store some key frame info on the crash record directly
Since shrinking the InnoDB pool size the stats page has
been very slow to load due to scanning the entire crash
and frame tables. Instead, store the bits it needs pre-
computed on the crash table and use that for all-time
top crashers calculation.

Will expand this out to other bits in the future.